refactor: rewrite features
